preserve float data for `strokeWidth` values on iOS
was formerly truncating due to `int` type

+++ ios/RNSketchCanvas/RNSketchCanvas/RNSketchCanvas.h
@@ -10,8 +10,8 @@
 
 - (BOOL)openSketchFile:(NSString *)filename directory:(NSString*) directory contentMode:(NSString*)mode;
 - (void)setCanvasText:(NSArray *)text;
-- (void)newPath:(int) pathId strokeColor:(UIColor*) strokeColor strokeWidth:(int) strokeWidth;
-- (void)addPath:(int) pathId strokeColor:(UIColor*) strokeColor strokeWidth:(int) strokeWidth points:(NSArray*) points;
+- (void)newPath:(int) pathId strokeColor:(UIColor*) strokeColor strokeWidth:(CGFloat) strokeWidth;
+- (void)addPath:(int) pathId strokeColor:(UIColor*) strokeColor strokeWidth:(CGFloat) strokeWidth points:(NSArray*) points;
 - (void)deletePath:(int) pathId;
 - (void)addPointX: (float)x Y: (float)y;
 - (void)endPath;
